Fixed Thread Problems

+++ Dev/VideoChatClient/ClientStartup/ClientStartup.hpp
@@ -13,6 +13,6 @@ namespace ClientStartup
 	uint32_t send_client_hello(Socket& socket, const std::string& username);
 	uint32_t client_startup_menu(Socket& socket, const uint32_t uuid, const std::string& username);
 
-	constexpr uint32_t UNKNOWN_MEETING_ID = ~0;
-	constexpr uint32_t UNKNOWN_USER_ID = ~0;
+	constexpr uint32_t UNKNOWN_MEETING_ID = static_cast<uint32_t>(~0);
+	constexpr uint32_t UNKNOWN_USER_ID = static_cast<uint32_t>(~0);
 }
